worker/pyproject.toml
[project]
name = "worker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "grpcio>=1.80.0",
    "grpcio-tools>=1.80.0",
    "mypy-protobuf>=5.0.0",
    "numpy>=2.4.4",
    "opencv-python>=4.13.0.92",
    "torch>=2.10.0",
    "ultralytics>=8.4.37",
]
